requirements.txt
fastapi~=0.115.12
python-multipart
pyjwt~=2.10.1
passlib[bcrypt]~=1.7.4
dotenv~=0.9.9
pydantic~=2.11.3
starlette~=0.46.2
python-dotenv~=1.1.0